changed spot seeder

+++ backend/db/migrations/20240105000210-create-spot.js
@@ -41,11 +41,11 @@ module.exports = {
         allowNull: false
       },
       lat: {
-        type: Sequelize.DECIMAL(3, 7),
+        type: Sequelize.FLOAT,
         allowNull: false
       },
       lng: {
-        type: Sequelize.DECIMAL(3, 7),
+        type: Sequelize.FLOAT,
         allowNull: false
       },
       name: {
